pyproject.toml
[tool.poetry]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = ["Paulo Roberto Moreira <prmoreiraf.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
django = ">=5.1.6,<6.0.0"
djangorestframework = "^3.16.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
